externals: move spack.util.ordereddict to external/ordereddict_backport
